Change transaction structure to JSON for list support
This update represents multi input output transactions as lists. All non unique values in the "vin" and "vout" dictionaries will now be stored in lists, with the index referring to which input or output they come from, respectively. For example, if a transaction has multiple inputs, the vin_txid column will now contain something along the lines of [vintxid1, vintxid2, vintxid3, ...] for any given row. This significantly simplifies the dataset and allows for more intuitive manipulation.