Changed names for simpler code and readability
PostgreSQL doesnt allow camel case without weird quoting - this simplifies the code. I plan to not drop indexes during the initial copy of populating the database for simplicity. Most users will not have a full node, so a slow copy is okay in order to keep up with the node.